Uncomment, add comments and rename module
