+++ src/index.js
@@ -1,5 +1,6 @@
 import parseOptions from './options';
 import prepare from './prepare/';
+import render from './render/';
 import parse from './parse/';
 import write from './write/';
 
@@ -10,7 +11,7 @@ import write from './write/';
  */
 function drizzle (options) {
   const opts = parseOptions(options);
-  return prepare(opts).then(parse).then(write);
+  return prepare(opts).then(parse).then(render).then(write);
 }
 
 export default drizzle;
